화면 맞추기

+++ teamProject/src/service/user/UserServiceImpl.java
@@ -1,11 +1,11 @@
 package service.user;
 
 import java.sql.SQLException;
-import java.util.List;
 
 import dao.user.UserDAO;
 import dao.user.UserDAOImpl;
 import dto.Customer;
+import session.SessionSet;
 
 public class UserServiceImpl implements UserService {
 
@@ -36,6 +36,8 @@ public class UserServiceImpl implements UserService {
 	 */
 	@Override
 	public void deleteMyInform(String id, String password) throws SQLException {
+		SessionSet ss  = SessionSet.getInstance();
+		if(!ss.get("user").getAttribute("id").equals(id))throw new SQLException("      로그인한 정보와 일치하지 않습니다. 다시 확인해주세요.");
 		int result = userDAO.deleteMyInform(id, password);
 		if(result!=0)throw new SQLException("        입력하신 정보에 일치하는 회원정보가 없습니다. 다시 한번 확인해주세요.");
 	}
